requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
streamlit
streamlit-chat
openvino-dev==2023.2.0
transformers==4.32.0
git+https://github.com/openvinotoolkit/nncf.git
sentencepiece
onnx>=1.11.0
torch==2.1.0
accelerate
tiktoken
transformers_stream_generator
einops
bitsandbytes